Don't mutate wire arrays while iterating.

+++ examples/functioncharts/functioncharts.ts
@@ -590,6 +590,7 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
     }
     function validateAndUpdate() {
       if (!self.isValidFunctionchart()) {
+        // TODO some kind of error message.
         self.transactionManager.cancelTransaction();
       }
       update();
@@ -763,12 +764,12 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
   forInWires(element: ElementTypes, visitor: WireVisitor) {
     element.inWires.forEach(wire => {
       if (wire)
-      visitor(wire);
+        visitor(wire);
     });
   }
 
   forOutWires(element: ElementTypes, visitor: WireVisitor) {
-    element.outWires.forEach((wires: Wire[]) => {
+    element.outWires.forEach(wires => {
       wires.forEach(wire => visitor(wire))
     });
   }
@@ -963,12 +964,24 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
     selection.set(roots.reverse());
   }
 
+  disconnectElement(element: ElementBase) {
+    const self = this;
+    element.inWires.forEach(wire => {
+      if (wire)
+        self.deleteItem(wire)
+    });
+    element.outWires.forEach(wires => {
+      if (wires.length === 0) return;
+      // Copy array since we're mutating it.
+      wires.splice(0).forEach(wire => {
+        if (wire)
+          self.deleteItem(wire);
+      });
+    });
+  }
   disconnectSelection() {
     const self = this;
-    this.selectedAllElements().forEach(element => {
-      self.forInWires(element, wire => this.deleteItem(wire));
-      self.forOutWires(element, wire => this.deleteItem(wire));
-    });
+    this.selectedAllElements().forEach(element => self.disconnectElement(element));
   }
 
   selectInteriorWires() {
@@ -1146,16 +1159,6 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
     });
   }
 
-  disconnectElement(element: ElementTypes, pin: number) {
-    element.inWires.forEach((wire, i) => {
-      if (wire)
-        this.deleteItem(wire);
-    });
-    element.outWires.forEach((wires, i) => {
-      wires.forEach(wire => this.deleteItem(wire));
-    });
-  }
-
   isValidWire(wire: Wire) {
     const src = wire.src,
           dst = wire.dst;
@@ -1241,7 +1244,7 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
         if (item.template.typeName === 'apply') {
           const type = self.resolveInputType(item, 0, new Set<ElementTypes>());
           let typeString = '[*,]';
-          if (type) {  // TODO clean up
+          if (type) {
             const newType = type.copy();
             newType.inputs.splice(0, 0, new Pin(Type.starType));
             newType.outputs.splice(0, 0, new Pin(Type.spacerType));
@@ -1270,18 +1273,8 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
       }
     });
 
-    // Delete dangling wires.
-    const invalidWires = new Array<Wire>();
-    this.wires.forEach(wire => {
-      if (!wire.src || !wire.dst || !self.elements.has(wire.src) || !self.elements.has(wire.dst))
-        invalidWires.push(wire);
-    });
-    invalidWires.forEach(wire => self.deleteItem(wire));
-
-    // Make sure wires between elements are in lowest common parent functionchart.
-    this.wires.forEach(wire => {
-      if (!self.isValidWire(wire))
-        invalidWires.push(wire);
+    // Make sure wires between elements are contained by the lowest common parent functionchart.
+    Array.from(this.wires.values()).forEach(wire => {   // Make a copy as we may mutate the set of wires.
       const src = wire.src!,
             dst = wire.dst!,
             srcParent = src.parent!,
@@ -1312,7 +1305,8 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
     // are deleted.
     const srcChange = new Array<Wire>(),
           dstChange = new Array<Wire>();
-    this.forInWires(element, wire => {
+    element.inWires.forEach(wire => {
+      if (!wire) return;
       const src = wire.src!, srcPin = wire.srcPin, dstPin = wire.dstPin;
       if (dstPin < newType.inputs.length &&
           src.type.outputs[srcPin].type.canConnectTo(newElement.type.inputs[dstPin].type)) {
@@ -1321,14 +1315,19 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
         this.deleteItem(wire);
       }
     });
-    this.forOutWires(element, wire => {
-      const dst = wire.dst!, srcPin = wire.srcPin, dstPin = wire.dstPin;
-      if (srcPin < newType.outputs.length &&
-        newElement.type.outputs[srcPin].type.canConnectTo(dst.type.inputs[dstPin].type)) {
-        srcChange.push(wire);
-      } else {
-        this.deleteItem(wire);
-      }
+    element.outWires.forEach(wires => {
+      if (wires.length === 0) return;
+      // Copy array since we're mutating.
+      wires.splice(0).forEach(wire => {
+        if (!wire) return;
+        const dst = wire.dst!, srcPin = wire.srcPin, dstPin = wire.dstPin;
+        if (srcPin < newType.outputs.length &&
+          newElement.type.outputs[srcPin].type.canConnectTo(dst.type.inputs[dstPin].type)) {
+          srcChange.push(wire);
+        } else {
+          this.deleteItem(wire);
+        }
+      });
     });
     srcChange.forEach(wire => {
       wire.src = newElement;
@@ -1339,12 +1338,6 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
 
     this.deleteItem(element);
   }
-
-  // updateType(instance: FunctionInstance) {
-  //   const newInstance = this.newFunctionInstance();
-  //   newInstance.functionchart = instance.functionchart;
-  //   this.replaceElement(instance, newInstance);
-  // }
 
   exportElement(element: Element | FunctionInstance) : Element {
     const result = this.newElement('export'),
@@ -1634,7 +1627,8 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
       // outWires.length >= outputs.
       for (let i = 0; i < outWires.length; i++) {
         const wires = outWires[i];
-        wires.forEach(wire => {
+        if (wires.length === 0) continue;
+        wires.splice(0).forEach(wire => {
           if (i >= outputs) {  // no pin at this index.
             self.deleteItem(wire);
           } else {
@@ -1715,7 +1709,8 @@ export class FunctionchartContext extends EventBase<Change, ChangeEvents>
           srcPin = wire.srcPin,
           dst = wire.dst,
           dstPin = wire.dstPin;
-    if (src && srcPin >= 0 && srcPin < src.type.outputs.length) {  // TODO how to deal with temporarily invalid wire?
+    if (src && srcPin >= 0 && srcPin < src.type.outputs.length) {
+      // TODO how to deal with temporarily invalid wire?
       const outputs = src.outWires[srcPin];
       if (!outputs.includes(wire))
         outputs.push(wire);
@@ -2358,7 +2353,7 @@ class Renderer {
   }
 
   hitTestWire(wire: Wire, p: Point, tol: number, mode: RenderMode) : WireHitResult | undefined {
-    // TODO don't hit test new wire as it's dragged!
+    // TODO don't hit test new wire as it's dragged.
     const hitInfo = hitTestBezier(wire.bezier, p, tol);
     if (hitInfo) {
       return new WireHitResult(wire, hitInfo);
